+++ src/SidePanel/TransmitSetupView.tsx
@@ -49,8 +49,8 @@ const TxPowerView = () => {
             }
             value={txPower}
             range={{
-                min: Math.min(...dbmValues),
-                max: Math.max(...dbmValues),
+                min: dbmValues.min,
+                max: dbmValues.max,
                 step: 1,
             }}
             disabled={isRunning}